+++ app.js
@@ -68,7 +68,8 @@ nwl.init({
   dirname,
   usernameField: "nickname",
   usersCollection: "user",
-  projectsCollection: "project"
+  projectsCollection: "project",
+  annotationsCollection: "mri"
 });
 global.authTokenMiddleware = nwl.authTokenMiddleware;
 const db = app.db.mongoDB();
@@ -150,193 +151,9 @@ atlasmakerServer.server.listen(8080, () => {
 });
 
 //========================================================================================
-// Admin route
+// Setup routes
 //========================================================================================
-app.use('/admin', require('./controller/admin/'));
-
-//========================================================================================
-// GUI routes
-//========================================================================================
-app.get('/', (req, res) => {
-  const login = (req.isAuthenticated()) ?
-    ('<a href=\'/user/' + req.user.username + '\'>' + req.user.username + '</a> (<a href=\'/logout\'>Log Out</a>)') :
-    ('<a href=\'/auth/github\'>Log in with GitHub</a>');
-
-  // store return path in case of login
-  req.session.returnTo = req.originalUrl;
-
-  res.render('index', {
-    title: 'BrainBox',
-    login: login
-  });
-});
-
-app.use('/mri', require('./controller/mri/'));
-app.use('/project', require('./controller/project/'));
-app.use('/user', require('./controller/user/'));
-
-//========================================================================================
-// API routes
-//========================================================================================
-app.get('/api/getLabelsets', (req, res) => {
-  const arr = fs.readdirSync(dirname + '/public/labels/');
-  const info = [];
-  for (const label of arr) {
-    var json = JSON.parse(fs.readFileSync(dirname + "/public/labels/" + label));
-    info.push({
-      name: json.name,
-      source: label
-    });
-  }
-  res.send(info);
-});
-
-app.get('/api/userNameQuery', (req, res) => {
-  const {query} = req;
-  db.get('user')
-    .find(
-      { $or: [
-        {nickname: {$regex:query.q}},
-        {name: {$regex:query.q}}
-      ]},
-      { fields: ['name', 'nickname'], limit: 10 }
-    )
-    .then((list) => {
-      res.send(list);
-    })
-    .catch(tracer.log);
-});
-
-app.get('/api/getAtlasBackups', (req, res) => {
-  const { source, atlasProject, atlasName } = req.query;
-
-  if(typeof source === "undefined"
-        || typeof atlasProject === "undefined"
-        || atlasName === "undefined") {
-    res.status(400);
-    res.render('error', {
-      message: "Missing source, atlasProject or atlasName"
-    });
-
-    return;
-  }
-
-  // get the mri object to which this atlas belongs
-  db.get('mri').findOne({
-    source: source,
-    "mri.atlas": {$elemMatch:{name: atlasName, project: atlasProject}},
-    backup: {$exists: 0}
-  }, {url: 1, "mri.atlas.$": 1})
-    .then( (obj) => {
-      // get all filenames that have ever been associated with this atlas
-      let {url: dataDir} = obj;
-      [,, dataDir] = dataDir.split("/");
-      db.get('mri').aggregate([
-        { $match:{ source: source, "mri.atlas":{$elemMatch: {project: atlasProject, name: atlasName}}}},
-        { $unwind: "$mri.atlas" },
-        { $match: { "mri.atlas.project":atlasProject, "mri.atlas.name": atlasName}},
-        { $group: {_id:{filename: "$mri.atlas.filename"}}},
-        { $project: {_id:0, filename:"$_id.filename"}}
-      ])
-        .then( (obj2) => {
-          // get all backups for those files...
-          let i;
-          const promiseArray = [];
-          // ...from backup logs
-          for(i=0; i<obj2.length; i++) {
-            promiseArray.push(
-              db.get('log').aggregate([
-                { $match: {key:"saveAtlasBackup", "value.atlasDirectory": dataDir, "value.atlasFilename": obj2[i].filename}},
-                { $project: {_id:0, filename:"$value.atlasFilename", timestamp:"$value.timestamp"}}
-              ])
-            );
-          }
-          Promise.all(promiseArray)
-            .then((values) => {
-              let result = [].concat(...values);
-              result = result.concat(obj2);
-              res.send(result);
-            })
-            .catch( (err) => {
-              res.status(500);
-              res.render('error', {
-                message: "Can't query backup file logs",
-                error: err
-              });
-            });
-        })
-        .catch( (err) => {
-          res.status(500);
-          res.render('error', {
-            message: "Can't query backup files",
-            error: err
-          });
-        });
-    })
-    .catch( (err) => {
-      res.status(400);
-      res.render('error', {
-        message: "Can't find atlas",
-        error: err
-      });
-    });
-});
-
-app.post('/api/log', (req, res) => {
-  const loggedUser = req.isAuthenticated() ? req.user.username : 'anonymous';
-  const json = req.body;
-  let obj;
-
-  switch (json.key) {
-  case 'annotationLength':
-    obj = {
-      key: "annotationLength",
-      username: loggedUser,
-      "value.source": json.value.source,
-      "value.atlas": json.value.atlas
-    };
-    req.db.get('log').findOne(obj)
-      .then( (result) => {
-        let length = 0;
-        if (result) {
-          length = parseFloat(result.value.length);
-        }
-        var sum = parseFloat(json.value.length) + length;
-        req.db.get('log').update(obj, {$set:{
-          "value.length":sum,
-          date: (new Date()).toJSON()
-        }}, {upsert: true});
-        res.send({length: sum});
-      })
-      .catch( (err) => {
-        tracer.log('ERROR', err);
-        res.send({error: JSON.stringify(err)});
-      });
-    break;
-  default:
-    req.db.get('log').insert({
-      key: json.key,
-      value: json.value,
-      username: loggedUser,
-      date: (new Date()).toJSON(),
-      ip: req.headers['x-forwarded-for'] ||
-                    req.connection.remoteAddress ||
-                    req.socket.remoteAddress ||
-                    req.connection.socket.remoteAddress
-    });
-    res.send();
-  }
-
-  req.db.get('mri').update({
-    source: json.value.source,
-    "mri.atlas":{$elemMatch:{filename:json.value.atlas}}
-  }, {
-    $set: {
-      "mri.atlas.$.modified": (new Date()).toJSON(),
-      "mri.atlas.$.modifiedBy": loggedUser
-    }
-  });
-});
+require('./controller/routes/routes')(app);
 
 //========================================================================================
 // Error handlers
